Refactor: extracted mobile keyboard handling logic
